+++ src/jsoncons/json2.hpp
@@ -362,8 +362,8 @@ void basic_json<Char, Alloc>::to_stream(basic_json_output_handler<Char>& handler
     case value_types::integer_t:
         handler.value(var_.value_.integer_value_);
         break;
-    case value_types::unsigned_integer_t:
-        handler.value(var_.value_.unsigned_integer_value_);
+    case value_types::uinteger_t:
+        handler.value(var_.value_.uinteger_value_);
         break;
     case value_types::bool_t:
         handler.value(var_.value_.bool_value_);
@@ -649,7 +649,7 @@ basic_json<Char, Alloc> basic_json<Char, Alloc>::parse_file(const std::string& f
 }
 
 template<typename Char, typename Alloc>
-typename basic_json<Char, Alloc>::object_iterator basic_json<Char, Alloc>::begin_members()
+typename basic_json<Char, Alloc>::member_iterator basic_json<Char, Alloc>::begin_members()
 {
     switch (var_.type_)
     {
@@ -678,7 +678,7 @@ typename basic_json<Char, Alloc>::const_object_iterator basic_json<Char, Alloc>:
 }
 
 template<typename Char, typename Alloc>
-typename basic_json<Char, Alloc>::object_iterator basic_json<Char, Alloc>::end_members()
+typename basic_json<Char, Alloc>::member_iterator basic_json<Char, Alloc>::end_members()
 {
     switch (var_.type_)
     {
@@ -839,8 +839,8 @@ long long basic_json<Char, Alloc>::as_longlong() const
         return static_cast<long long>(var_.value_.float_value_);
     case value_types::integer_t:
         return static_cast<long long>(var_.value_.integer_value_);
-    case value_types::unsigned_integer_t:
-        return static_cast<long long>(var_.value_.unsigned_integer_value_);
+    case value_types::uinteger_t:
+        return static_cast<long long>(var_.value_.uinteger_value_);
     case value_types::bool_t:
         return var_.value_.bool_value_ ? 1 : 0;
     default:
@@ -857,8 +857,8 @@ unsigned long long basic_json<Char, Alloc>::as_ulonglong() const
         return static_cast<unsigned long long>(var_.value_.float_value_);
     case value_types::integer_t:
         return static_cast<unsigned long long>(var_.value_.integer_value_);
-    case value_types::unsigned_integer_t:
-        return static_cast<unsigned long long>(var_.value_.unsigned_integer_value_);
+    case value_types::uinteger_t:
+        return static_cast<unsigned long long>(var_.value_.uinteger_value_);
     case value_types::bool_t:
         return var_.value_.bool_value_ ? 1 : 0;
     default:
@@ -875,8 +875,8 @@ double basic_json<Char, Alloc>::as_double() const
         return var_.value_.float_value_;
     case value_types::integer_t:
         return static_cast<double>(var_.value_.integer_value_);
-    case value_types::unsigned_integer_t:
-        return static_cast<double>(var_.value_.unsigned_integer_value_);
+    case value_types::uinteger_t:
+        return static_cast<double>(var_.value_.uinteger_value_);
     case value_types::null_t:
         return std::numeric_limits<double>::quiet_NaN();
     default:
@@ -893,8 +893,8 @@ int basic_json<Char, Alloc>::as_int() const
         return static_cast<int>(var_.value_.float_value_);
     case value_types::integer_t:
         return static_cast<int>(var_.value_.integer_value_);
-    case value_types::unsigned_integer_t:
-        return static_cast<int>(var_.value_.unsigned_integer_value_);
+    case value_types::uinteger_t:
+        return static_cast<int>(var_.value_.uinteger_value_);
     case value_types::bool_t:
         return var_.value_.bool_value_ ? 1 : 0;
     default:
@@ -911,8 +911,8 @@ unsigned int basic_json<Char, Alloc>::as_uint() const
         return static_cast<unsigned int>(var_.value_.float_value_);
     case value_types::integer_t:
         return static_cast<unsigned int>(var_.value_.integer_value_);
-    case value_types::unsigned_integer_t:
-        return static_cast<unsigned int>(var_.value_.unsigned_integer_value_);
+    case value_types::uinteger_t:
+        return static_cast<unsigned int>(var_.value_.uinteger_value_);
     case value_types::bool_t:
         return var_.value_.bool_value_ ? 1 : 0;
     default:
@@ -929,8 +929,8 @@ long basic_json<Char, Alloc>::as_long() const
         return static_cast<long>(var_.value_.float_value_);
     case value_types::integer_t:
         return static_cast<long>(var_.value_.integer_value_);
-    case value_types::unsigned_integer_t:
-        return static_cast<long>(var_.value_.unsigned_integer_value_);
+    case value_types::uinteger_t:
+        return static_cast<long>(var_.value_.uinteger_value_);
     case value_types::bool_t:
         return var_.value_.bool_value_ ? 1 : 0;
     default:
@@ -947,8 +947,8 @@ unsigned long basic_json<Char, Alloc>::as_ulong() const
         return static_cast<unsigned long>(var_.value_.float_value_);
     case value_types::integer_t:
         return static_cast<unsigned long>(var_.value_.integer_value_);
-    case value_types::unsigned_integer_t:
-        return static_cast<unsigned long>(var_.value_.unsigned_integer_value_);
+    case value_types::uinteger_t:
+        return static_cast<unsigned long>(var_.value_.uinteger_value_);
     case value_types::bool_t:
         return var_.value_.bool_value_ ? 1 : 0;
     default: