+++ src/jsoncons/json_deserializer.hpp
@@ -194,7 +194,7 @@ private:
         }
     }
 
-    void do_unsigned_integer_value(uint64_t value, const basic_parsing_context<Char>&) override
+    void do_uinteger_value(uint64_t value, const basic_parsing_context<Char>&) override
     {
         if (top_ == -1)
         {