+++ src/jsoncons/json_input_handler.hpp
@@ -13,6 +13,55 @@
 
 namespace jsoncons {
 
+template<typename CharT>
+uint64_t string_to_uinteger(const CharT *s, size_t length) throw(std::overflow_error)
+{
+    static const uint64_t max_value = std::numeric_limits<uint64_t>::max JSONCONS_NO_MACRO_EXP();
+    static const uint64_t max_value_div_10 = max_value / 10;
+    uint64_t n = 0;
+    for (size_t i = 0; i < length; ++i)
+    {
+        uint64_t x = s[i] - '0';
+        if (n > max_value_div_10)
+        {
+            throw std::overflow_error("Unsigned overflow");
+        }
+        n = n * 10;
+        if (n > max_value - x)
+        {
+            throw std::overflow_error("Unsigned overflow");
+        }
+
+        n += x;
+    }
+    return n;
+}
+
+template<typename CharT>
+int64_t string_to_integer(bool has_neg, const CharT *s, size_t length) throw(std::overflow_error)
+{
+    const long long max_value = std::numeric_limits<int64_t>::max JSONCONS_NO_MACRO_EXP();
+    const long long max_value_div_10 = max_value / 10;
+
+    long long n = 0;
+    for (size_t i = 0; i < length; ++i)
+    {
+        long long x = s[i] - '0';
+        if (n > max_value_div_10)
+        {
+            throw std::overflow_error("Integer overflow");
+        }
+        n = n * 10;
+        if (n > max_value - x)
+        {
+            throw std::overflow_error("Integer overflow");
+        }
+
+        n += x;
+    }
+    return has_neg ? -n : n;
+}
+
 template <typename Char>
 class basic_parsing_context;
 
@@ -94,17 +143,17 @@ public:
 
     void value(unsigned int value, const basic_parsing_context<Char>& context) 
     {
-        do_unsigned_integer_value(value,context);
+        do_uinteger_value(value,context);
     }
 
     void value(unsigned long value, const basic_parsing_context<Char>& context) 
     {
-        do_unsigned_integer_value(value,context);
+        do_uinteger_value(value,context);
     }
 
     void value(unsigned long long value, const basic_parsing_context<Char>& context) 
     {
-        do_unsigned_integer_value(value,context);
+        do_uinteger_value(value,context);
     }
 
     void value(float value, const basic_parsing_context<Char>& context)
@@ -150,7 +199,7 @@ private:
 
     virtual void do_integer_value(int64_t value, const basic_parsing_context<Char>& context) = 0;
 
-    virtual void do_unsigned_integer_value(uint64_t value, const basic_parsing_context<Char>& context) = 0;
+    virtual void do_uinteger_value(uint64_t value, const basic_parsing_context<Char>& context) = 0;
 
     virtual void do_bool_value(bool value, const basic_parsing_context<Char>& context) = 0;
 };
@@ -210,7 +259,7 @@ private:
     {
     }
 
-    void do_unsigned_integer_value(unsigned long long, const basic_parsing_context<Char>&) override
+    void do_uinteger_value(unsigned long long, const basic_parsing_context<Char>&) override
     {
     }